💄 fix(compaction): group tool rows inside the fold rail
Folded turns rendered their children individually, losing the tool-row grouping
the live transcript uses. Extract groupRenderItems into tool-call-group.tsx
(shared, no import cycle) and apply it inside the fold rail so runs of tool
calls collapse into a ToolCallGroup there too.
