+++ frontend/src/components/chat-view.tsx
@@ -54,7 +54,7 @@ import {
   sessionHasKnowledgeChanges,
 } from "@/lib/utils";
 import { Message } from "./message";
-import { ToolCallGroup } from "./tool-call-group";
+import { ToolCallGroup, groupRenderItems } from "./tool-call-group";
 import { ChatInput } from "./chat-input";
 import { AgentHistoryView } from "./agent-history-view";
 
@@ -312,56 +312,6 @@ function argsMatch(
 }
 
 /** A run of consecutive tool/thinking messages, collapsed into one summary. */
-type RenderItem =
-  | { type: "message"; index: number }
-  | { type: "group"; start: number; indices: number[]; inProgress: boolean };
-
-function isToolish(m: ChatMessage): boolean {
-  return (
-    m.kind === "tool_call" ||
-    m.kind === "thinking" ||
-    m.kind === "thinking_streaming"
-  );
-}
-
-/**
- * Group maximal runs of tool/thinking messages. A run becomes a collapsible
- * group only when it has ≥2 items and contains ≥1 tool_call; otherwise its
- * items render individually. A group that reaches the end of the list is still
- * in progress (rendered expanded).
- */
-function groupRenderItems(messages: ChatMessage[]): RenderItem[] {
-  const items: RenderItem[] = [];
-  let i = 0;
-  while (i < messages.length) {
-    if (!isToolish(messages[i])) {
-      items.push({ type: "message", index: i });
-      i++;
-      continue;
-    }
-    const start = i;
-    const indices: number[] = [];
-    while (i < messages.length && isToolish(messages[i])) {
-      indices.push(i);
-      i++;
-    }
-    const hasToolCall = indices.some(
-      (j) => messages[j].kind === "tool_call",
-    );
-    if (indices.length >= 2 && hasToolCall) {
-      items.push({
-        type: "group",
-        start,
-        indices,
-        inProgress: i >= messages.length,
-      });
-    } else {
-      for (const j of indices) items.push({ type: "message", index: j });
-    }
-  }
-  return items;
-}
-
 function applyDeniedApprovalToMessages(
   messages: ChatMessage[],
   request: {
